+++ src/main/java/tornadofx/Component.kt
@@ -900,7 +900,6 @@ abstract class UIComponent(viewTitle: String? = "", icon: Node? = null) : Compon
     @JvmOverloads
     fun openModal(stageStyle: StageStyle = StageStyle.DECORATED, modality: Modality = Modality.APPLICATION_MODAL, escapeClosesWindow: Boolean = true, owner: Window? = currentWindow, block: Boolean = false, resizable: Boolean? = null): Stage? {
         if (modalStage == null) {
-            require(getRootWrapper() is Parent) { "Only Parent Fragments can be opened in a Modal" }
             modalStage = Stage(stageStyle)
             // modalStage needs to be set before this code to make close() work in blocking mode
             with(modalStage!!) {
@@ -910,18 +909,18 @@ abstract class UIComponent(viewTitle: String? = "", icon: Node? = null) : Compon
                 initModality(modality)
                 if (owner != null) initOwner(owner)
 
+                if (escapeClosesWindow) {
+                    addEventFilter(KeyEvent.KEY_PRESSED) {
+                        if (it.code == KeyCode.ESCAPE)
+                            close()
+                    }
+                }
+
                 if (getRootWrapper().scene != null) {
                     scene = getRootWrapper().scene
                     this@UIComponent.properties["tornadofx.scene"] = getRootWrapper().scene
                 } else {
                     Scene(getRootWrapper()).apply {
-                        if (escapeClosesWindow) {
-                            addEventFilter(KeyEvent.KEY_PRESSED) {
-                                if (it.code == KeyCode.ESCAPE)
-                                    close()
-                            }
-                        }
-
                         FX.applyStylesheetsTo(this)
                         val primaryStage = FX.getPrimaryStage(scope)
                         if (primaryStage != null) icons += primaryStage.icons