+++ src/main/java/tornadofx/InternalWindow.kt
@@ -164,7 +164,7 @@ class InternalWindow(icon: Node?, modal: Boolean, escapeClosesWindow: Boolean, c
 
         children.add(0, owner)
         fillOverlay()
-        view.callOnDock()
+//        view.callOnDock()
     }
 
     fun close() {