+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - `onCancel()` is now called when reusing Wizard instance 
 - Observable collection delegates removed because they shadow observable properties (See Properties.kt:L212)
+- The `style` property of `ListCell` will be cleared by the framework so it can be manipulated in `cellFormat` without side effects (https://stackoverflow.com/questions/51459371/custom-cell-format-listview-tornadofx-on-delete-item)
+- Escape closes window only works first time for Views (https://github.com/edvin/tornadofx/issues/764)
 
 ### Changes
 
@@ -22,6 +24,7 @@
 - `finally(callback)` for runAsync and other task builders
 - `Window.aboutToBeShown` property avoid false positives for invisible Workspace warning (https://github.com/edvin/tornadofx/issues/755)
 - Slideshow slides supports optional timeout, which will advance to the next slide using the Slide transition
+- `importStylesheet` now supports `file`, `http` and `https` in addition to classpath resources (https://github.com/edvin/tornadofx/issues/762)
 
 ## [1.7.16]