+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - SortedFilteredList.setAllPassThrough property controls if `setAll` should be forwarded to the underlying list (https://github.com/edvin/tornadofx/issues/344) and (https://github.com/edvin/tornadofx/issues/681)
 - EventBus deliveries will continue even after a subscriber throws exception
 - UIComponent callback `onBeforeShow` is called for all top level UIComponents
+- Primary View `onDock` is now called *after* the stage is shown. Now it aligns with the timing for secondary windows.
 
 ### Additions