+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- TreeView.`lazypopulate` leafcheck logic was reversed (https://github.com/edvin/tornadofx/issues/773)
 - `onCancel()` is now called when reusing Wizard instance 
 - Observable collection delegates removed because they shadow observable properties (See Properties.kt:L212)
 - The `style` property of `ListCell` will be cleared by the framework so it can be manipulated in `cellFormat` without side effects (https://stackoverflow.com/questions/51459371/custom-cell-format-listview-tornadofx-on-delete-item)